Add database finding
add database function find, search and filter

+++ Lesson06/src/main/java/com/example/Lesson06/services/ProductService.java
@@ -11,5 +11,5 @@ public interface ProductService {
 
     Product save(Product product);
 
-    List<Product> getByPrice(Float startFilter, Float endFilter);
+    List<Product> getByPrice(Double startFilter, Double endFilter);
 }
